Move Wwise hierarchy and related classes to own file
