+++ audio_modder.py
@@ -11,6 +11,7 @@ import wave
 import sys
 import pathlib
 import xml.etree.ElementTree as etree
+from wwise_hierarchy import WwiseHierarchy
 
 from functools import partial
 from functools import cmp_to_key
@@ -533,273 +534,7 @@ class MediaIndex:
         arr = [x.get_data() for x in self.entries.values()]
         data_arr = self.data.values()
         return b"".join(arr) + b"".join(data_arr)
-                
-class HircEntry:
-    
-    def __init__(self):
-        self.size = self.hierarchy_type = self.hierarchy_id = self.misc = 0
-        self.sources = []
-        self.track_info = []
-        self.soundbank = None
-        self.modified = False
-    
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = HircEntry()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        entry.hierarchy_id = stream.uint32_read()
-        entry.misc = stream.read(entry.size - 4)
-        return entry
-        
-    @classmethod
-    def from_bytes(cls, data):
-        stream = MemoryStream()
-        stream.write(data)
-        stream.seek(0)
-        return cls.from_memory_stream(stream)
-        
-    def import_entry(self, new_entry):
-        pass
-        
-    def get_id(self):
-        return self.hierarchy_id
-        
-    def raise_modified(self):
-        self.soundbank.raise_modified()
-        
-    def lower_modified(self):
-        self.soundbank.lower_modified()
-        
-    def get_data(self):
-        return self.hierarchy_type.to_bytes(1, byteorder="little") + self.size.to_bytes(4, byteorder="little") + self.hierarchy_id.to_bytes(4, byteorder="little") + self.misc
-        
-class MusicRandomSequence(HircEntry):
-    
-    def __init__(self):
-        super().__init__()
-    
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = MusicRandomSequence()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        entry.hierarchy_id = stream.uint32_read()
-        return entry
-        
-    def get_data(self):
-        pass
-        
-class RandomSequenceContainer(HircEntry):
-    def __init__(self):
-        super().__init__()
-        self.unused_sections = []
-        self.contents = []
-        
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = RandomSequenceContainer()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        start_position = stream.tell()
-        entry.hierarchy_id = stream.uint32_read()
-
-        # ---------------------------------------
-        section_start = stream.tell()
-        stream.advance(1)
-        n = stream.uint8_read() #num fx
-        if n == 0:
-            stream.advance(12)
-        else:
-            stream.advance(7*n + 13)
-        stream.advance(5*stream.uint8_read()) #number of props
-        stream.advance(9*stream.uint8_read()) #number of props (again)
-        if stream.uint8_read() & 0b0000_0010: #positioning bit vector
-            if stream.uint8_read() & 0b0100_0000: # relative pathing bit vector
-                stream.advance(5)
-                stream.advance(16*stream.uint32_read())
-                stream.advance(20*stream.uint32_read())
-        if stream.uint8_read() & 0b0000_1000: #I forget what this is for
-            stream.advance(26)
-        else:
-           stream.advance(10)
-        stream.advance(3*stream.uint8_read()) #num state props
-        for _ in range(stream.uint8_read()): #num state groups
-            stream.advance(5)
-            stream.advance(8*stream.uint8_read())
-        for _ in range(stream.uint16_read()):  # num RTPC
-            stream.advance(12)
-            stream.advance(stream.uint16_read()*12)
-        section_end = stream.tell()
-        # ---------------------------------------
-
-        stream.seek(section_start)
-        entry.unused_sections.append(stream.read(section_end-section_start+24))
-
-        for _ in range(stream.uint32_read()): #number of children (tracks)
-            entry.contents.append(stream.uint32_read())
-
-        entry.unused_sections.append(stream.read(entry.size - (stream.tell()-start_position)))
-        return entry
-        
-    def get_data(self):
-        return (
-            b"".join([
-                struct.pack("<BII", self.hierarchy_type, self.size, self.hierarchy_id),
-                self.unused_sections[0],
-                len(self.contents).to_bytes(4, byteorder="little"),
-                b"".join([x.to_bytes(4, byteorder="little") for x in self.contents]),
-                self.unused_sections[1]
-            ])
-        )
-    
-class MusicSegment(HircEntry):
-
-    def __init__(self):
-        super().__init__()
-        self.tracks = []
-        self.duration = 0
-        self.entry_marker = None
-        self.exit_marker = None
-        self.unused_sections = []
-        self.markers = []
-        self.modified = False
-    
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = MusicSegment()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        entry.hierarchy_id = stream.uint32_read()
-        entry.unused_sections.append(stream.read(15))
-        n = stream.uint8_read() #number of props
-        stream.seek(stream.tell()-1)
-        entry.unused_sections.append(stream.read(5*n + 1))
-        n = stream.uint8_read() #number of props (again)
-        stream.seek(stream.tell()-1)
-        entry.unused_sections.append(stream.read(5*n + 1 + 12 + 4)) #the 4 is the count of state props, state chunks, and RTPC, which are currently always 0
-        n = stream.uint32_read() #number of children (tracks)
-        for _ in range(n):
-            entry.tracks.append(stream.uint32_read())
-        entry.unused_sections.append(stream.read(23)) #meter info
-        n = stream.uint32_read() #number of stingers
-        stream.seek(stream.tell()-4)
-        entry.unused_sections.append(stream.read(24*n + 4))
-        entry.duration = struct.unpack("<d", stream.read(8))[0]
-        n = stream.uint32_read() #number of markers
-        for i in range(n):
-            id = stream.uint32_read()
-            position = struct.unpack("<d", stream.read(8))[0]
-            name = []
-            temp = b"1"
-            while temp != b"\x00":
-                temp = stream.read(1)
-                name.append(temp)
-            name = b"".join(name)
-            marker = [id, position, name]
-            entry.markers.append(marker)
-            if i == 0:
-                entry.entry_marker = marker
-            elif i == n-1:
-                entry.exit_marker = marker
-        return entry
-        
-    def set_data(self, duration=None, entry_marker=None, exit_marker=None):
-        if not self.modified:
-            self.duration_old = self.duration
-            self.entry_marker_old = self.entry_marker[1]
-            self.exit_marker_old = self.exit_marker[1]
-            self.raise_modified()
-        if duration is not None: self.duration = duration
-        if entry_marker is not None: self.entry_marker[1] = entry_marker
-        if exit_marker is not None: self.exit_marker[1] = exit_marker
-        self.modified = True
-        
-    def import_entry(self, new_entry):
-        if self.modified:
-            pass
-        
-    def revert_modifications(self):
-        if self.modified:
-            self.lower_modified()
-            self.entry_marker[1] = self.entry_marker_old
-            self.exit_marker[1] = self.exit_marker_old
-            self.duration = self.duration_old
-            self.modified = False
-        
-    def get_data(self):
-        return (
-            b"".join([
-                struct.pack("<BII", self.hierarchy_type, self.size, self.hierarchy_id),
-                self.unused_sections[0],
-                self.unused_sections[1],
-                self.unused_sections[2],
-                len(self.tracks).to_bytes(4, byteorder="little"),
-                b"".join([x.to_bytes(4, byteorder="little") for x in self.tracks]),
-                self.unused_sections[3],
-                self.unused_sections[4],
-                struct.pack("<d", self.duration),
-                len(self.markers).to_bytes(4, byteorder="little"),
-                b"".join([b"".join([x[0].to_bytes(4, byteorder="little"), struct.pack("<d", x[1]), x[2]]) for x in self.markers])
-            ])
-        )
-        
-class HircEntryFactory:
-    
-    @classmethod
-    def from_memory_stream(cls, stream):
-        hierarchy_type = stream.uint8_read()
-        stream.seek(stream.tell()-1)
-        if hierarchy_type == 2: #sound
-            return Sound.from_memory_stream(stream)
-        elif hierarchy_type == 11: #music track
-            return MusicTrack.from_memory_stream(stream)
-        elif hierarchy_type == 0x0A: #music segment
-            return MusicSegment.from_memory_stream(stream)
-        elif hierarchy_type == 0x05: #random sequence container
-            return RandomSequenceContainer.from_memory_stream(stream)
-        else:
-            return HircEntry.from_memory_stream(stream)
-        
-class WwiseHierarchy:
-    
-    def __init__(self, soundbank = None):
-        self.entries = {}
-        self.type_lists = {}
-        self.soundbank = soundbank
-        
-    def load(self, hierarchy_data):
-        self.entries.clear()
-        reader = MemoryStream()
-        reader.write(hierarchy_data)
-        reader.seek(0)
-        num_items = reader.uint32_read()
-        for item in range(num_items):
-            entry = HircEntryFactory.from_memory_stream(reader)
-            entry.soundbank = self.soundbank
-            self.entries[entry.get_id()] = entry
-            try:
-                self.type_lists[entry.hierarchy_type].append(entry)
-            except:
-                self.type_lists[entry.hierarchy_type] = [entry]
-                
-    def import_hierarchy(self, new_hierarchy):
-        for entry in new_hierarchy.get_entries():
-            self.get_entry(entry.get_id()).import_entry(entry)
-                   
-    def get_entry(self, entry_id):
-        return self.entries[entry_id]
-        
-    def get_entries(self):
-        return self.entries.values()
-        
-    def get_type(self, hirc_type):
-        return self.types[hirc_type]
-            
-    def get_data(self):
-        arr = [entry.get_data() for entry in self.entries.values()]
-        return len(arr).to_bytes(4, byteorder="little") + b"".join(arr)
-            
+                         
 class BankParser:
     
     def __init__(self):
@@ -824,126 +559,7 @@ class BankParser:
             return self.chunks[chunk_tag]
         except:
             return None
-            
-class BankSourceStruct:
 
-    def __init__(self):
-        self.plugin_id = 0
-        self.stream_type = self.source_id = self.mem_size = self.bit_flags = 0
-        
-    @classmethod
-    def from_bytes(cls, bytes):
-        b = BankSourceStruct()
-        b.plugin_id, b.stream_type, b.source_id, b.mem_size, b.bit_flags = struct.unpack("<IBIIB", bytes)
-        return b
-        
-    def get_data(self):
-        return struct.pack("<IBIIB", self.plugin_id, self.stream_type, self.source_id, self.mem_size, self.bit_flags)
-        
-class TrackInfoStruct:
-    
-    def __init__(self):
-        self.track_id = self.source_id = self.event_id = self.play_at = self.begin_trim_offset = self.end_trim_offset = self.source_duration = 0
-        self.play_at_old = self.begin_trim_offset_old = self.end_trim_offset_old = self.source_duration_old = 0
-        self.modified = False
-        self.soundbanks = set()
-        
-    @classmethod
-    def from_bytes(cls, bytes):
-        t = TrackInfoStruct()
-        t.track_id, t.source_id, t.event_id, t.play_at, t.begin_trim_offset, t.end_trim_offset, t.source_duration = struct.unpack("<IIIdddd", bytes)
-        return t
-        
-    def get_id(self):
-        if self.source_id != 0:
-            return self.source_id
-        else:
-            return self.event_id
-            
-    def is_modified(self):
-        return self.modified
-            
-    def set_data(self, play_at=None, begin_trim_offset=None, end_trim_offset=None, source_duration=None):
-        if not self.modified:
-            self.play_at_old = self.play_at
-            self.begin_trim_offset_old = self.begin_trim_offset
-            self.end_trim_offset_old = self.end_trim_offset
-            self.source_duration_old = self.source_duration
-            self.raise_modified()
-        if play_at is not None: self.play_at = play_at
-        if begin_trim_offset is not None: self.begin_trim_offset = begin_trim_offset
-        if end_trim_offset is not None: self.end_trim_offset = end_trim_offset
-        if source_duration is not None: self.source_duration = source_duration
-        self.modified = True
-        
-    def revert_modifications(self):
-        if self.modified:
-            self.lower_modified()
-            self.play_at = self.play_at_old
-            self.begin_trim_offset = self.begin_trim_offset_old
-            self.end_trim_offset = self.end_trim_offset_old
-            self.source_duration = self.source_duration_old
-            self.modified = False
-            
-    def raise_modified(self):
-        for bank in self.soundbanks:
-            bank.raise_modified()
-        
-    def lower_modified(self):
-        for bank in self.soundbanks:
-            bank.lower_modified()
-        
-    def get_data(self):
-        return struct.pack("<IIIdddd", self.track_id, self.source_id, self.event_id, self.play_at, self.begin_trim_offset, self.end_trim_offset, self.source_duration)
-            
-class MusicTrack(HircEntry):
-    
-    def __init__(self):
-        super().__init__()
-        self.bit_flags = 0
-        
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = MusicTrack()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        start_position = stream.tell()
-        entry.hierarchy_id = stream.uint32_read()
-        entry.bit_flags = stream.uint8_read()
-        num_sources = stream.uint32_read()
-        for _ in range(num_sources):
-            source = BankSourceStruct.from_bytes(stream.read(14))
-            entry.sources.append(source)
-        num_track_info = stream.uint32_read()
-        for _ in range(num_track_info):
-            track = TrackInfoStruct.from_bytes(stream.read(44))
-            entry.track_info.append(track)
-        entry.misc = stream.read(entry.size - (stream.tell()-start_position))
-        return entry
-
-    def get_data(self):
-        b = b"".join([source.get_data() for source in self.sources])
-        t = b"".join([track.get_data() for track in self.track_info])
-        return struct.pack("<BIIBI", self.hierarchy_type, self.size, self.hierarchy_id, self.bit_flags, len(self.sources)) + b + len(self.track_info).to_bytes(4, byteorder="little") + t + self.misc
-    
-class Sound(HircEntry):
-    
-    def __init__(self):
-        super().__init__()
-    
-    @classmethod
-    def from_memory_stream(cls, stream):
-        entry = Sound()
-        entry.hierarchy_type = stream.uint8_read()
-        entry.size = stream.uint32_read()
-        entry.hierarchy_id = stream.uint32_read()
-        entry.sources.append(BankSourceStruct.from_bytes(stream.read(14)))
-        entry.misc = stream.read(entry.size - 18)
-        return entry
-
-    def get_data(self):
-        return struct.pack(f"<BII14s{len(self.misc)}s", self.hierarchy_type, self.size, self.hierarchy_id, self.sources[0].get_data(), self.misc)
-        
 class WwiseBank(Subscriber):
     
     def __init__(self):
@@ -1014,7 +630,7 @@ class WwiseBank(Subscriber):
     def get_data(self):
         return self.data
             
-    def generate(self, audio_sources, eventTrackInfo):
+    def generate(self, audio_sources):
         data = bytearray()
         data += self.bank_header
         
@@ -1040,10 +656,7 @@ class WwiseBank(Subscriber):
         
         added_sources = set()
         
-        for entry in self.hierarchy.entries.values():
-            for index, info in enumerate(entry.track_info):
-                if info.event_id != 0:
-                    entry.track_info[index] = eventTrackInfo[info.event_id]
+        for entry in self.hierarchy.get_type(SOUND):
             for source in entry.sources:
                 bank_generation_progress_window.step()
                 if source.plugin_id == VORBIS:
@@ -1051,16 +664,6 @@ class WwiseBank(Subscriber):
                         audio = audio_sources[source.source_id]
                     except KeyError:
                         continue
-                    try:
-                        count = 0
-                        for info in entry.track_info:
-                            if info.source_id == source.source_id:
-                                break
-                            count += 1
-                        if audio.get_track_info() is not None: #is this needed?
-                            entry.track_info[count] = audio.get_track_info()
-                    except: #exception because there may be no original track info struct
-                        pass
                     if source.stream_type == PREFETCH_STREAM and source.source_id not in added_sources:
                         data_array.append(audio.get_data()[:source.mem_size])
                         didx_array.append(struct.pack("<III", source.source_id, offset, source.mem_size))
@@ -1357,7 +960,7 @@ class FileReader:
             bank = self.wwise_banks[key]
             toc_file.write(bank.toc_header.get_data())
             toc_file.seek(bank.toc_header.toc_data_offset)
-            toc_file.write(pad_to_16_byte_align(bank.toc_data_header + bank.generate()))
+            toc_file.write(pad_to_16_byte_align(bank.toc_data_header + bank.generate(self.audio_sources)))
             
         for key in self.wwise_banks.keys():
             toc_file.seek(file_position)
